# apps/api/src/routes/cases.ts
import type { FastifyInstance } from "fastify";
import { z } from "zod";

import {
  buildErrorResponse,
  caseCreateSchema,
  Errors,
  walletAttestationSchema
} from "@kadenatrace/shared";

import type { CaseService } from "../services/case-service.js";

const walletSignerSchema = z.object({
  accountName: z.string().min(3),
  publicKey: z.string().min(16),
  adapterName: z.string().min(1).optional()
});

const signedCommandSchema = z.object({
  cmd: z.string().min(3),
  hash: z.string().min(10),
  sigs: z.array(z.any())
});

const anchorPrepareSchema = z.object({
  signer: walletSignerSchema
});

const anchorSubmitSchema = z.object({
  signer: walletSignerSchema,
  signedCommand: signedCommandSchema
});

const attestationDraftSchema = z.object({
  wallet: z.string().min(3),
  chain: z.enum(["ethereum", "bsc", "kadena", "bitcoin"]),
  riskLevel: z.enum(["low", "medium", "high", "critical"]),
  riskScore: z.number().min(0).max(100),
  note: z.string().max(500).optional()
});

const attestationPrepareSchema = z.object({
  signer: walletSignerSchema,
  attestation: attestationDraftSchema
});

const attestationSubmitSchema = z.object({
  signer: walletSignerSchema,
  attestation: attestationDraftSchema,
  signedCommand: signedCommandSchema
});

const disputeCreateSchema = z.object({
  reason: z.string().min(10).max(1000),
  signer: walletSignerSchema
});

const disputeSubmitSchema = z.object({
  disputeId: z.string().min(3),
  signedCommand: signedCommandSchema
});

export async function registerCaseRoutes(app: FastifyInstance, caseService: CaseService) {
  app.get("/api/cases/by-chain/:chain", async (request, reply) => {
    try {
      const params = request.params as { chain: string };
      const query = request.query as { cursor?: string; limit?: string };

      const result = await caseService.listCasesByChain(params.chain, {
        cursor: query.cursor,
        limit: query.limit ? parseInt(query.limit, 10) : undefined
      });

      return reply.send(result);
    } catch (error) {
      request.log.error(error);
      const { statusCode, body } = buildErrorResponse(error);
      return reply.code(statusCode).send(body);
    }
  });

  app.post("/api/cases", async (request, reply) => {
    const parsed = caseCreateSchema.safeParse(request.body);
    if (!parsed.success) {
      const error = Errors.validationError(
        "body",
        "Invalid case creation request",
        parsed.error.flatten()
      );
      const { statusCode, body } = buildErrorResponse(error);
      return reply.code(statusCode).send(body);
    }

    try {
      const record = await caseService.createCase(parsed.data);
      return reply.send({
        caseId: record.caseId,
        slug: record.slug,
        publicUri: record.publicUri,
        traceHash: record.traceHash,
        verifiable: true
      });
    } catch (error) {
      request.log.error(error);
      const { statusCode, body } = buildErrorResponse(error);
      return reply.code(statusCode).send(body);
    }
  });

  app.post("/api/cases/:caseId/anchor/payload", async (request, reply) => {
    const params = request.params as { caseId: string };
    const parsed = anchorPrepareSchema.safeParse(request.body);
    if (!parsed.success) {
      const error = Errors.validationError(
        "body",
        "Invalid anchor prepare request",
        parsed.error.flatten()
      );
      const { statusCode, body } = buildErrorResponse(error);
      return reply.code(statusCode).send(body);
    }

    try {
      const payload = await caseService.prepareCaseAnchor(params.caseId, parsed.data.signer);
      return reply.send(payload);
    } catch (error) {
      request.log.error(error);
      const { statusCode, body } = buildErrorResponse(error);
      return reply.code(statusCode).send(body);
    }
  });

  app.post("/api/cases/:caseId/anchor", async (request, reply) => {
    const params = request.params as { caseId: string };
    try {
      const record = await caseService.anchorCase(params.caseId);
      return reply.send(record.anchor);
    } catch (error) {
      request.log.error(error);
      const { statusCode, body } = buildErrorResponse(error);
      return reply.code(statusCode).send(body);
    }
  });

  app.post("/api/cases/:caseId/anchor/submit", async (request, reply) => {
    const params = request.params as { caseId: string };
    const parsed = anchorSubmitSchema.safeParse(request.body);
    if (!parsed.success) {
      const error = Errors.validationError(
        "body",
        "Invalid anchor submit request",
        parsed.error.flatten()
      );
      const { statusCode, body } = buildErrorResponse(error);
      return reply.code(statusCode).send(body);
    }

    try {
      const record = await caseService.submitCaseAnchor(params.caseId, parsed.data.signer, parsed.data.signedCommand);
      return reply.send(record.anchor);
    } catch (error) {
      request.log.error(error);
      const { statusCode, body } = buildErrorResponse(error);
      return reply.code(statusCode).send(body);
    }
  });

  app.post("/api/cases/:caseId/attestations", async (request, reply) => {
    const params = request.params as { caseId: string };
    const parsed = walletAttestationSchema.safeParse(request.body);
    if (!parsed.success) {
      const error = Errors.validationError(
        "body",
        "Invalid attestation request",
        parsed.error.flatten()
      );
      const { statusCode, body } = buildErrorResponse(error);
      return reply.code(statusCode).send(body);
    }

    try {
      const record = await caseService.addAttestation(params.caseId, parsed.data);
      return reply.send({ caseId: record.caseId, attestations: record.attestations });
    } catch (error) {
      request.log.error(error);
      const { statusCode, body } = buildErrorResponse(error);
      return reply.code(statusCode).send(body);
    }
  });

  app.post("/api/cases/:caseId/attestations/payload", async (request, reply) => {
    const params = request.params as { caseId: string };
    const parsed = attestationPrepareSchema.safeParse(request.body);
    if (!parsed.success) {
      const error = Errors.validationError(
        "body",
        "Invalid attestation prepare request",
        parsed.error.flatten()
      );
      const { statusCode, body } = buildErrorResponse(error);
      return reply.code(statusCode).send(body);
    }

    try {
      const payload = await caseService.prepareWalletAttestation(
        params.caseId,
        parsed.data.attestation,
        parsed.data.signer
      );
      return reply.send(payload);
    } catch (error) {
      request.log.error(error);
      const { statusCode, body } = buildErrorResponse(error);
      return reply.code(statusCode).send(body);
    }
  });

  app.post("/api/cases/:caseId/attestations/submit", async (request, reply) => {
    const params = request.params as { caseId: string };
    const parsed = attestationSubmitSchema.safeParse(request.body);
    if (!parsed.success) {
      const error = Errors.validationError(
        "body",
        "Invalid attestation submit request",
        parsed.error.flatten()
      );
      const { statusCode, body } = buildErrorResponse(error);
      return reply.code(statusCode).send(body);
    }

    try {
      const record = await caseService.submitWalletAttestation(
        params.caseId,
        parsed.data.attestation,
        parsed.data.signer,
        parsed.data.signedCommand
      );
      return reply.send({ caseId: record.caseId, attestations: record.attestations });
    } catch (error) {
      request.log.error(error);
      const { statusCode, body } = buildErrorResponse(error);
      return reply.code(statusCode).send(body);
    }
  });

  // New dispute endpoints
  app.post("/api/cases/:caseId/disputes/payload", async (request, reply) => {
    const params = request.params as { caseId: string };
    const parsed = disputeCreateSchema.safeParse(request.body);
    if (!parsed.success) {
      const error = Errors.validationError(
        "body",
        "Invalid dispute request",
        parsed.error.flatten()
      );
      const { statusCode, body } = buildErrorResponse(error);
      return reply.code(statusCode).send(body);
    }

    try {
      const payload = await caseService.prepareDispute(params.caseId, parsed.data.reason, parsed.data.signer);
      return reply.send(payload);
    } catch (error) {
      request.log.error(error);
      const { statusCode, body } = buildErrorResponse(error);
      return reply.code(statusCode).send(body);
    }
  });

  app.post("/api/cases/:caseId/disputes/submit", async (request, reply) => {
    const params = request.params as { caseId: string };
    const parsed = disputeSubmitSchema.safeParse(request.body);
    if (!parsed.success) {
      const error = Errors.validationError(
        "body",
        "Invalid dispute submit request",
        parsed.error.flatten()
      );
      const { statusCode, body } = buildErrorResponse(error);
      return reply.code(statusCode).send(body);
    }

    try {
      const record = await caseService.submitDispute(params.caseId, parsed.data.disputeId, parsed.data.signedCommand);
      return reply.send(record);
    } catch (error) {
      request.log.error(error);
      const { statusCode, body } = buildErrorResponse(error);
      return reply.code(statusCode).send(body);
    }
  });

  app.get("/api/cases/:caseId/disputes", async (request, reply) => {
    const params = request.params as { caseId: string };
    try {
      const disputes = await caseService.listDisputesForCase(params.caseId);
      return reply.send(disputes);
    } catch (error) {
      request.log.error(error);
      const { statusCode, body } = buildErrorResponse(error);
      return reply.code(statusCode).send(body);
    }
  });

  app.get("/api/public/cases", async (request, reply) => {
    try {
      const query = request.query as { cursor?: string; limit?: string };
      const result = await caseService.listPublicCases({
        cursor: query.cursor,
        limit: query.limit ? parseInt(query.limit, 10) : undefined
      });
      return reply.send(result);
    } catch (error) {
      request.log.error(error);
      const { statusCode, body } = buildErrorResponse(error);
      return reply.code(statusCode).send(body);
    }
  });

  app.get("/api/cases/:caseId/attestations", async (request, reply) => {
    try {
      const params = request.params as { caseId: string };
      const query = request.query as { cursor?: string; limit?: string };
      const result = await caseService.listAttestationsForCase(params.caseId, {
        cursor: query.cursor,
        limit: query.limit ? parseInt(query.limit, 10) : undefined
      });
      return reply.send(result);
    } catch (error) {
      request.log.error(error);
      const { statusCode, body } = buildErrorResponse(error);
      return reply.code(statusCode).send(body);
    }
  });

  app.get("/api/public/cases/:slug", async (request, reply) => {
    const params = request.params as { slug: string };
    const record = await caseService.getPublicCase(params.slug);
    if (!record) {
      const error = Errors.caseNotFound(params.slug);
      const { statusCode, body } = buildErrorResponse(error);
      return reply.code(statusCode).send(body);
    }

    return reply.send(record);
  });
}